// src/common/enums.ts
export enum VerifyCredentials {
  verify_email = 'email',
  verify_password = 'password',
}

export enum SqlType {
  SQL = 'SQL',
  NORMAL = 'NORMAL',
  BR = 'BR',
}

export enum SortingType {
  ID = 'ID',
  NAME = 'NAME',
  DATE = 'DATE',
  NUMBER = 'NUMBER',
  FIRST_DATE = 'FIRST_DATE',
  LAST_DATE = 'LAST_DATE',
}

export enum ShiftType {
  PRIMEIRO_TURNO = '1° Turno',
  SEGUNDO_TURNO = '2° Turno',
}

export enum ValidType {
  IS_BOOLEAN = 'IS_BOOLEAN',
  NO_SPACE = 'NO_SPACE',
  NO_MANY_SPACE = 'NO_MANY_SPACE',
  IS_STRING = 'IS_STRING',
  IS_NUMBER = 'IS_NUMBER',
  IS_EXCEL_NUMBER = 'IS_EXCEL_NUMBER',
  IS_NUMBER_FLOAT = 'IS_NUMBER_FLOAT',
  NO_SPECIAL_CHARACTER = 'NO_SPECIAL_CHARACTER',
  OFFICE_NO_SPECIAL_CHARACTER = 'OFFICE_NO_SPECIAL_CHARACTER',
  SHIFT_NO_SPECIAL_CHARACTER = 'SHIFT_NO_SPECIAL_CHARACTER',
  IS_EMAIL = 'IS_EMAIL',
  DATE = 'DATE',
  DATE_BR = 'DATE_BR',
  IS_CNPJ = 'IS_CNPJ',
  SPECIAL_CHARACTER = 'SPECIAL_CHARACTER',
  MINIMUM_ONE_NUMBER = 'MINIMUM_ONE_NUMBER',
  MINIMUM_ONE_NUMBER_STRING_SPECIAL_CHARACTER = 'MINIMUM_ONE_NUMBER_STRING_SPECIAL_CHARACTER',
  MINIMUM_ONE_STRING = 'MINIMUM_ONE_STRING',
  DIFERENT_OF_ZERO = 'DIFERENT_OF_ZERO',
  SLOTS_BLISTER = 'SLOTS_BLISTERS',
  BLANK_SPACES_BEGINNING = 'BLANK_SPACES_BEGINNING',
  MST_DIFERENT_OF_ZERO = 'MST_DIFERENT_OF_ZERO',
  MST_IS_NUMBER = 'MST_IS_NUMBER',
  CABINET_DIFERENT_OF_ZERO = 'CABINET_DIFERENT_OF_ZERO',
  CABINET_IS_NUMBER = 'CABINET_IS_NUMBER',
  IS_PIPE_CODE = 'IS_PIPE_CODE',
  IS_PIPE_NAME = 'IS_PIPE_NAME',
  IS_MACHINE_IP = 'IS_MACHINE_IP',
}

export enum ObjectSize {
  INTEGER = 2147483646,
  DEFAULT_DAYS = 35,
}

export enum DateOperation {
  ADD = 'ADD',
  REMOVE = 'REMOVE',
}

export enum TypeEmail {
  CREATE = 'CREATE',
  RESET = 'RESET',
}

export enum CodeError {
  IS_REGISTERED = 1000,
  NOT_FOUND = 1001,
  INVALID_NUMBER = 1002,
  NOT_EMPTY = 1003,
  IS_DESABLE = 1004,
  IS_BLOCKED = 1005,
  CANNOT_BE_CHANGED = 1006,
  INCORRECT_FIRST_ACCESS_PASSWORD = 1101,
  INCORRECT_PASSWORD_ENROLLMENT = 1111,
  NEW_PASSWORD_CANNOT_MATCH = 1102,
  PASSWORD_CONFIRMATION_NOT_MATCH = 1103,
  IS_BIT = 1200,
  IS_NUMBER = 1201,
  IS_STRING = 1202,
  NO_SPACE = 1203,
  NO_SPECIAL_CHARACTER = 1204,
  IS_EMAIL = 1205,
  SPECIAL_CHARACTER = 1206,
  MIN_ONE_NUMBER = 1207,
  MIN_CHARACTER = 1208,
  MAX_CHARACTER = 1209,
  DIFERENT_OF_ZERO = 1210,
  IS_NOT_BOOLEAN = 1211,
  NO_ZERO = 1212,
  IS_NOT_NUMBER = 1213,
  NO_MANY_SPACE = 1214,
  INVALID_DATE = 1215,
  MIN_ONE_NUMBER_STRING_SPECIAL_CHARACTER = 1216,
  IS_INTEGER = 1217,
  MAX_SIZE = 1218,
  INVALID_TYPE = 1219,
  INVALID_DATE_RANGER = 1220,
  EMPTY_FILE = 1221,
  BLANK_SPACES_BEGINNING = 2424,
  INVALID_DATA_RANGER = 3000,
  IN_PROCESS = 4000,
  REENVIADO = 5000,
  NOT_ALLOWED_RESTRICTED_TRANSACTIONS_ONLY = 6000,
  MAX_ONE_TRANSACTION_FOR_RESOURCE = 6001,
  NOT_ALLOWED = 6002,
  NOT_ALLOWED_SAME_TRANSACTIONS = 6003,
  IS_PIPE_CODE = 7000,
  IS_MACHINE_IP = 8000,
  EXTERNAL_LOGIN_ERROR = 9000,
}

export enum CodeObject {
  NAME = 'Name',
  ENROLLMENT = 'Enrollment',
  PASSWORD = 'Password',
  USER = 'User',
  SHIFT = 'Shift',
  EMAIL = 'Email',
  PROFILE = 'Profile',
  ID = 'Id',
  IMAGE = 'Image',
  VALUE = 'Value',
}

export enum RestrictTransactions {
  USER_RESTRICT = 1,
  DASHBOARD_RESTRICT = 9,
  REPORT_RESTRICT = 13,
  MODEL_RESTRICT = 24,
}

export enum UserTransactions {
  USER_RESTRICT = 1,
  USER_VIEW = 2,
  USER_EDIT = 3,
  USER_TOTAL = 4,
}

export enum DashboardTransactions {
  DASHBOARD_RESTRICT = 9,
  DASHBOARD_VIEW = 10,
  DASHBOARD_EDIT = 11,
  DASHBOARD_TOTAL = 12,
}

export enum ReportTransactions {
  REPORT_RESTRICT = 13,
  REPORT_VIEW = 14,
  REPORT_EDIT = 15,
  REPORT_TOTAL = 16,
}

export enum ModelTransactions {
  MODEL_RESTRICT = 24,
  MODEL_VIEW = 25,
  MODEL_EDIT = 26,
  MODEL_TOTAL = 27,
}